move error.go to config package
